REAFCT: Support inheritance
https://github.com/51Degrees/ip-intelligence-dotnet/issues/72
https://github.com/51Degrees/device-detection-dotnet/pull/399
